Support removing cached files (#132)
* Add empty demo app

* Display fake image

* Store response type for fake requests

* Create directory for images

* Remove hardcoded logic

* Add SwiftFormat to demo project

* Add a way to clear downloads folder

* Wrap up

* Move test

* Rename test method

* Rename

+++ Sources/NSFileManager+Helpers.swift
@@ -10,6 +10,8 @@ extension FileManager {
 
     public func remove(at url: URL) {
         let path = url.path
+        guard FileManager.default.isDeletableFile(atPath: url.path) else { return }
+
         do {
             try FileManager.default.removeItem(atPath: path)
         } catch let error as NSError {
